Added Account.cs , ConsumerApp.cs and updated stringliteral.cs and ConsumerAppRepository.cs. Fixes #144
