+++ Promact.Oauth.Server/src/Promact.Oauth.Server/StringLiterals/Account.cs
@@ -1,0 +1,8 @@
+﻿namespace Promact.Oauth.Server.StringLiterals
+{
+    public class Account
+    {
+        public string EmailNotExists { get; set; }
+        public string SuccessfullySendMail { get; set; }
+    }
+}
