+++ Promact.Oauth.Server/src/Promact.Oauth.Server/StringLiterals/ConsumerApp.cs
@@ -1,0 +1,8 @@
+﻿namespace Promact.Oauth.Server.StringLiterals
+{
+    public class ConsumerApp
+    {
+        public string AlphaNumericString { get; set; }
+        public string CapitalAlphaNumericString { get; set; }
+    }
+}
