--- Promact.Oauth.Server/src/Promact.Oauth.Server/StringLiterals/StringLiteralss.cs
@@ -1,25 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Threading.Tasks;
-
-namespace Promact.Oauth.Server.StringLiterals
-{
-    
-    public class ConsumerApp
-    {
-        public string ATOZaTOz0TO9 { get; set; }
-        public string ATOZ0TO9 { get; set; }
-    }
-    public class Account
-    {
-        public string EmailNotExists { get; set; }
-        public string SuccessfullySendMail { get; set; }
-    }
-    public class StringLiteral
-    {
-        public Account Account { get; set; }
-        public ConsumerApp ConsumerApp { get; set; }
-    }
-    
-}
