--- Promact.Oauth.Server/src/Promact.Oauth.Server/StringLiterals/stringliteral.json
@@ -1,14 +1,0 @@
-﻿{
-  "StringLiterals": {
-    "ConsumerApp": {
-      "AlphaNumericString": "ABCDEFGHIJKLMNOPQRSTUVWXYZabcdefghijklmnopqrstuvwxyz0123456789",
-      "CapitalAlphaNumericString": "ABCDEFGHIJKLMNOPQRSTUVWXYZ0123456789"
-
-    },
-    "Account": {
-      "EmailNotExists": "Email does not exist",
-      "SuccessfullySendMail": "We have sent you a link on {{emailaddress}} to reset password.Please check your email."
-    }
-   
-  }
-}
